+++ components/src/main/java/org/patternfly/component/list/DescriptionListTerm.java
@@ -20,6 +20,7 @@ import org.jboss.elemento.ElementTextDelegate;
 import org.patternfly.component.ComponentIcon;
 import org.patternfly.component.popover.Popover;
 import org.patternfly.style.Classes;
+
 import elemental2.dom.Element;
 import elemental2.dom.HTMLElement;