+++ components/src/main/java/org/patternfly/component/popover/PopperPopover.java
@@ -37,9 +37,9 @@ import org.patternfly.handler.CloseHandler;
 import org.patternfly.popper.Modifiers;
 import org.patternfly.popper.Popper;
 import org.patternfly.popper.PopperBuilder;
+import org.patternfly.popper.PopperPlacement;
 import org.patternfly.popper.TriggerAction;
 import org.patternfly.style.Modifiers.NoPadding;
-import org.patternfly.popper.PopperPlacement;
 
 import elemental2.dom.Element;
 import elemental2.dom.Event;
@@ -67,6 +67,8 @@ import static org.patternfly.core.Roles.dialog;
 import static org.patternfly.handler.CloseHandler.fireEvent;
 import static org.patternfly.handler.CloseHandler.shouldClose;
 import static org.patternfly.icon.IconSets.fas.times;
+import static org.patternfly.popper.PopperPlacement.auto;
+import static org.patternfly.popper.PopperPlacement.top;
 import static org.patternfly.style.Classes.arrow;
 import static org.patternfly.style.Classes.close;
 import static org.patternfly.style.Classes.component;
@@ -74,8 +76,6 @@ import static org.patternfly.style.Classes.content;
 import static org.patternfly.style.Classes.modifier;
 import static org.patternfly.style.Classes.popover;
 import static org.patternfly.style.Classes.widthAuto;
-import static org.patternfly.popper.PopperPlacement.auto;
-import static org.patternfly.popper.PopperPlacement.top;
 
 /**
  * A popover is in-app messaging that provides more information on specific product areas. Popovers display content in a new