+++ core/src/main/java/org/patternfly/popper/PopperBuilder.java
@@ -24,7 +24,6 @@ import java.util.function.Predicate;
 import org.gwtproject.event.shared.HandlerRegistration;
 import org.jboss.elemento.EventType;
 import org.jboss.elemento.logger.Logger;
-import org.patternfly.popper.PopperPlacement;
 
 import elemental2.core.JsArray;
 import elemental2.dom.CSSProperties;